// src/app/core/interfaces/contributions.ts
export interface ContributionsInterface {
  company: string;
  position: string;
  location: string;
  yearOfStart: number;
  link: string;
  active: boolean;
}

export interface StudiesInterface {
  school: string;
  career: string;
  location: string;
  yearOfStart: string;
  yearOfEnd?:string;
  concluded: boolean;
}